Chore: Moved utils into plugins and destructure import
